requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.9.220
botocore==1.12.220
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
chart-studio==1.0.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==2.2.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
docutils==0.15.2
gensim==3.8.0
graphviz==0.12
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
JPype1==0.7.0
kiwisolver==1.1.0
konlpy==0.5.1
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
nltk==3.4.5
numpy==1.17.1
openapi-codec==1.3.2
pandas==0.25.1
Pillow==6.1.0
plotly==4.1.1
psycopg2==2.8.3
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
scipy==1.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
smart-open==1.8.4
soupsieve==1.9.3
sqlparse==0.3.0
squarify==0.4.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wordcloud==1.5.0
wrapt==1.11.2
